+++ src/vgr/builtins/web.py
@@ -8,16 +8,18 @@ from urllib.error import URLError
 
 from .common import dist_x, str_arg
 from .types import poly_to_boolean
+from .registry import builtin
 
+@builtin("ParseUrl")
 def parse_url(url: Any=None, remove_nulls: bool=True) -> Any:
     """
-**Decompose a URL string and return a dictionary of its components**
+**Decompose a URL string into a dictionary of its components**
 
 * ParseUrl(*value*)
 * *value*.ParseUrl()
 
-In addition, the result will contain _error_ indicating if there was an error
-and _error_msg_ describing the error. It always include the URL itself.
+In addition, the result will contain *error* indicating if there was an error
+and *error_msg* describing the error. It always includes the URL itself.
 
 Distributive over lists but not dictionaries.
 
@@ -39,6 +41,7 @@ parsed.query → "x=1&y=2"
 parsed.fragment → "frag"
 ```
 """
+    def _empty_to_none(s:str) -> str: return None if s is None or len(s) == 0 else s
     if url is None: return None
     remove_nulls = poly_to_boolean(remove_nulls)
     if isinstance(url, list): return dist_x(parse_url, url, remove_nulls)
@@ -48,50 +51,51 @@ parsed.fragment → "frag"
     try:
         parsed = urlparse(url)
         components = {
-            "url": url,
-            "error": False,
-            "error_msg": None,
-            "netloc": _empty_to_none(parsed.netloc),
-            "scheme": _empty_to_none(parsed.scheme),
-            "hostname": _empty_to_none(parsed.hostname),
-            "path": _empty_to_none(parsed.path),
-            "params": _empty_to_none(parsed.params),
-            "query": _empty_to_none(parsed.query),
-            "fragment": _empty_to_none(parsed.fragment),
-            "port": parsed.port,
-            "username": _empty_to_none(parsed.username),
-            "password": _empty_to_none(parsed.password),
+            "url":        url,
+            "error":      False,
+            "error_msg":  None,
+            "netloc":    _empty_to_none(parsed.netloc),
+            "scheme":    _empty_to_none(parsed.scheme),
+            "hostname":  _empty_to_none(parsed.hostname),
+            "path":      _empty_to_none(parsed.path),
+            "params":    _empty_to_none(parsed.params),
+            "query":     _empty_to_none(parsed.query),
+            "fragment":  _empty_to_none(parsed.fragment),
+            "port":      parsed.port,
+            "username":  _empty_to_none(parsed.username),
+            "password":  _empty_to_none(parsed.password),
         }
         return {key: value for key, value in components.items() if value is not None} if remove_nulls else components
     except (ValueError, URLError, TypeError) as e:
         return {
-            "url": url,
-            "error": False,
+            "url":       url,
+            "error":     True,
             "error_msg": str(e),
         }
 
+@builtin("EncodeUrl")
 def encode_url(url: str=None, safe: str="/") -> str:
     """
 **Encode reserved characters in a full or partial URL**
 
-* EncodeURL(*value*)
-* EncodeURL(*value*, _safe_)
-* *value*.EncodeURL()
-* *value*.EncodeURL(_safe_)
+* EncodeUrl(*value*)
+* EncodeUrl(*value*, *safe*)
+* *value*.EncodeUrl()
+* *value*.EncodeUrl(*safe*)
 
-The _safe_ argument is string which lists reserved characters that need
+The *safe* argument is string which lists reserved characters that need
 not be encoded. When omitted, it defaults to "/".
 
 ```vgr
-None.EncodeURL() → None
-5.EncodeURL() → "5"
-"hello world".EncodeURL() → "hello%20world"
-"a/b/c".EncodeURL() → "a/b/c"
-"a/b/c".EncodeURL("") → "a%2Fb%2Fc"
-"role=admin".EncodeURL() → "role%3Dadmin"
-"Café Münster".EncodeURL() → "Caf%C3%A9%20M%C3%BCnster"
+None.EncodeUrl() → None
+5.EncodeUrl() → "5"
+"hello world".EncodeUrl() → "hello%20world"
+"a/b/c".EncodeUrl() → "a/b/c"
+"a/b/c".EncodeUrl("") → "a%2Fb%2Fc"
+"role=admin".EncodeUrl() → "role%3Dadmin"
+"Café Münster".EncodeUrl() → "Caf%C3%A9%20M%C3%BCnster"
 
-// NB: use care when quoting an entire URL
+// NB: use caution when quoting an entire URL
 "http://example.com?q=lst".EncodeURL() → "http%3A//example.com%3Fq%3Dlst"
 ```
 """
@@ -102,6 +106,3 @@ None.EncodeURL() → None
     if isinstance(url, (int, float)): url = str(url)
     if isinstance(url, str): return quote(url, safe=safe, errors='strict', encoding='utf-8')
     return url
-
-def _empty_to_none(s:str) -> str:
-    return None if s is None or len(s) == 0 else s