+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 
 ### Changed
 
+- Breaking change: Left/RightShift() renamed to ShiftLeft/Right()
+  to match RotateLeft/Right()
 - Constant statement can now create a constant from an initialized variable
   without having to use an assignment (e.g. "Set x To 5; Constant x").
   Additionally, this form of the statement is idempotent.